+++ src/wifi/helper/wifi-helper.h
@@ -13,6 +13,7 @@
 
 #include "wifi-mac-helper.h"
 
+#include "ns3/iana-link-type-numbers.h"
 #include "ns3/qos-utils.h"
 #include "ns3/trace-helper.h"
 #include "ns3/wifi-net-device.h"
@@ -163,11 +164,11 @@ class WifiPhyHelper : public PcapHelperForDevice, public AsciiTraceHelperForDevi
     enum SupportedPcapDataLinkTypes
     {
         DLT_IEEE802_11 =
-            PcapHelper::DLT_IEEE802_11, /**< IEEE 802.11 Wireless LAN headers on packets */
+            iana::LinkType::IEEE802_11, /**< IEEE 802.11 Wireless LAN headers on packets */
         DLT_PRISM_HEADER =
-            PcapHelper::DLT_PRISM_HEADER, /**< Include Prism monitor mode information */
+            iana::LinkType::IEEE802_11_PRISM, /**< Include Prism monitor mode information */
         DLT_IEEE802_11_RADIO =
-            PcapHelper::DLT_IEEE802_11_RADIO /**< Include Radiotap link layer information */
+            iana::LinkType::IEEE802_11_RADIOTAP /**< Include Radiotap link layer information */
     };
 
     /**
@@ -195,7 +196,7 @@ class WifiPhyHelper : public PcapHelperForDevice, public AsciiTraceHelperForDevi
          * @param dev the WifiNetDevice for which the PCAP files are generated
          */
         PcapFilesInfo(const std::string& filename,
-                      PcapHelper::DataLinkType dlt,
+                      iana::LinkType dlt,
                       WifiPhyHelper::PcapCaptureType type,
                       Ptr<WifiNetDevice> dev)
             : commonFilename{filename},
@@ -207,7 +208,7 @@ class WifiPhyHelper : public PcapHelperForDevice, public AsciiTraceHelperForDevi
         }
 
         std::string commonFilename;              ///< common file name of the PCAP files
-        PcapHelper::DataLinkType pcapDlt;        ///< the selected data link type of the pcap file
+        iana::LinkType pcapDlt;                  ///< the selected data link type of the pcap file
         WifiPhyHelper::PcapCaptureType pcapType; ///< the selected PCAP capture type
         Ptr<WifiNetDevice> device; ///< the WifiNetDevice for which the PCAP files are generated
         std::map<uint8_t, Ptr<PcapFileWrapper>> files; ///< PCAP files indexed by PHY ID
@@ -231,7 +232,7 @@ class WifiPhyHelper : public PcapHelperForDevice, public AsciiTraceHelperForDevi
      *
      * @returns The data link type of the pcap file (and packets) to be used
      */
-    PcapHelper::DataLinkType GetPcapDataLinkType() const;
+    iana::LinkType GetPcapDataLinkType() const;
 
     /**
      * Set the PCAP capture type to be used. This function has to be called before EnablePcap().
@@ -363,8 +364,8 @@ class WifiPhyHelper : public PcapHelperForDevice, public AsciiTraceHelperForDevi
                              Ptr<NetDevice> nd,
                              bool explicitFilename) override;
 
-    PcapHelper::DataLinkType m_pcapDlt; ///< PCAP data link type
-    PcapCaptureType m_pcapType;         ///< PCAP capture type
+    iana::LinkType m_pcapDlt;   ///< PCAP data link type
+    PcapCaptureType m_pcapType; ///< PCAP capture type
 };
 
 /**